+++ MainGUI/Resources/Changelog.md
@@ -3,6 +3,7 @@ Modnix Changelog
 # Version ?
 
 * Fix: Save, Enable, and Disable buttons are now disabled when app is busy or game is running.
+* Fix: When config panel can't show config, it will display an error instead of stuck on old tab.
 * Mod: New api "api_list" and "api_info".
 * Mod: "api_add" and "api_remove" now rejects if there are content after api key.